Added requirements_cpu.txt
It seems some of the versions of programs listed don't exist on pip. This was the closest I could get, and this file can be used to quickly install all requirements with `pip install -r requirements_cpu.txt`. This install works for me.

Note, I am running python 3.8.5 on macOS with a MacBook Pro, no GPU support. I think the package for torch might be different when using a GPU, so someone else that has a GPU can make that requirements file with `pip freeze > requirements_gpu.txt` after they get it running in a minimal virtual environment.